Remove @Mod annotation from main Vigilance class
This library should itself not be a mod, rather, it should be a utility
to be used by other mods. To that end, the main class has been modified
so that it does not register itself as a mod.

Now, mods that use Vigilance must call Vigilance.initialize() at least
once before using it.