Revert "Settings: Allow the value to be set externally"
This reverts commit a17f3e2f1cd2e73cb8c2cb6853498368fbf05a79. This
commit wasn't really ready to be pushed (I pushed it on accident). The
external setting of component values needs to be thought out a bit more,
and executed in a way where the components can be used externally (this
was not the case here, as I made them all accept PropertyData objects).