ColorPicker: Fix rendering when partially outside of visible area
Previously the components which do the colorful drawing were not the same size
as the things which they drew. This caused them to not draw at all even though
only their top part (which as far as the component is concerned, is everything)
was outside the visible area.

This commit changes the bounds of the components to match their content and also
cleans up some weird off-by-one-in-two-places conditions in the rendering.