Migrations: Use explicit `Migration` interface instead of fun type
So the Java people don't have to return Unit everywhere.
Also gives us the option to add further methods to `Migration`s in the future if
necessary.

Does have the downside of requiring Kotlin people to add an explicit `Migration`
before the lambda for SAM conversion. But since each lambda will usually be on
its own line anyway (unlike in the tests), that doesn't seem too bad.